Implements long term cloudtrail monitoring solution.
Hello @adunsulag and @jesdynf ! Hope this message finds you well. This PR implements a long term cloud trail logging solution. The trail logs all events in the same region that contains all the infrastructure for the installation and stores those logs in S3 for 7 years and in Cloudwatch logs for 9 years. It can be disabled for testing purposes but it's turned on by default. The KMS key we provision for it increases our minimum monthly cost by $1/month.